Split manage workspace overview panels
