+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -56,3 +56,4 @@
 | LWB-R051 | DPM rebalance-wave display policy extraction | Hardened | `dpm-wave-command-center-panel.tsx` still owned rebalance lifecycle step labels and badge-tone display policy inline, keeping pure source-state presentation rules inside the largest DPM panel. That increased review noise around Gateway-backed wave/campaign actions. | Moved rebalance lifecycle display steps and source-state badge-tone mapping into `dpm-wave-command-center-panel-helpers.ts`; updated the panel to consume those helpers without changing Gateway-only wave, campaign, proof-pack, or launch-history calls; added focused helper coverage for lifecycle labels and success/warn/danger/default tone mapping. | Continue extracting rebalance-wave rendering only when it reduces real review complexity. Keep campaign membership/readiness calculation, trade approval, order generation, routing, fills, settlement, OMS execution, maker-checker claims, and external execution claims out of Workbench unless source-owned services and control design explicitly back them. | `src/features/workbench/dpm-wave-command-center-panel-helpers.ts`, `src/features/workbench/components/dpm-wave-command-center-panel.tsx`, `tests/unit/dpm-wave-command-center-panel-helpers.test.ts`; validation via `npm run test -- --run tests/unit/dpm-wave-command-center-panel-helpers.test.ts`, `npm run test -- --run tests/unit/dpm-wave-command-center-panel.test.tsx`, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R052 | Outcome-review client boundary component extraction | Hardened | `outcome-review-panel.tsx` rendered the source-owned client-communication boundary inline after the downstream realization slice, mixing boundary display policy with Gateway-backed outcome-review actions. That made the no-client-communication posture harder to test independently and increased review noise in the main panel. | Moved the boundary subview into `outcome-review-client-boundary-card.tsx`; updated the panel to consume the component without changing Gateway-only supportability, report-input, or AI-evidence handoffs; added focused component coverage that proves the boundary renders from source-owned fields and exposes no client communication action controls. | Keep client messaging, approval, delivery, communication audit, and client-contact workflows out of Workbench unless a source-owning service exposes an explicit governed contract and control design. | `src/features/workbench/components/outcome-review-client-boundary-card.tsx`, `src/features/workbench/components/outcome-review-panel.tsx`, `tests/unit/outcome-review-client-boundary-card.test.tsx`; validation via `npm run test -- --run tests/unit/outcome-review-client-boundary-card.test.tsx tests/unit/outcome-review-panel.test.tsx`, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
 | LWB-R053 | Manage workspace orchestration module split | Hardened | `manage-workspace.tsx` owned Gateway data loading, source-owned identifier discovery, mode navigation policy, and React rendering in one large module. `manage-workspace-view-model.ts` also imported the data type from that UI file, creating avoidable coupling between pure view-model helpers and the page component. | Moved Manage data loading and source-owned id selectors into `manage-workspace-data.ts`; moved Manage mode definitions, href building, rail-item shaping, and mode normalization into `manage-workspace-navigation.ts`; updated the route/page, view-model, and mandate-health component to depend on the focused modules; added unit coverage for navigation policy and source-owned id selectors. | Continue reducing `manage-workspace.tsx` by extracting overview and context-rail subviews when the next slice can preserve the advisor workflow and avoid introducing unsupported Manage capability. | `src/features/workbench/manage-workspace-data.ts`, `src/features/workbench/manage-workspace-navigation.ts`, `src/features/workbench/manage-workspace.tsx`, `src/app/workbench/[portfolioId]/page.tsx`, `tests/unit/manage-workspace-data.test.ts`, `tests/unit/manage-workspace-navigation.test.ts`; validation via `npm run test -- --run tests/unit/manage-workspace-data.test.ts tests/unit/manage-workspace-navigation.test.ts tests/integration/workbench-page.test.tsx`, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
+| LWB-R054 | Manage overview and context rail component split | Hardened | After the orchestration split, `manage-workspace.tsx` still owned the full Manage overview and side-rail rendering. That kept advisor workflow presentation, rail next-action policy, and mode orchestration coupled in one file, making future Manage screen changes noisier than necessary. | Moved the overview surface into `manage-overview.tsx` and the advisor context rail into `manage-context-rail.tsx`; reduced `manage-workspace.tsx` to shell, mode dispatch, and unavailable-state orchestration; added component tests that render the split surfaces directly from Gateway-shaped Manage data and assert no client-communication action is exposed. | Continue extracting smaller Manage overview subviews only if the overview grows again. Keep mandate readiness, PM quality evidence, proof-pack posture, outcome evidence, and client communication truth source-owned and Gateway-backed. | `src/features/workbench/components/manage-overview.tsx`, `src/features/workbench/components/manage-context-rail.tsx`, `src/features/workbench/manage-workspace.tsx`, `tests/unit/manage-workspace-components.test.tsx`; validation via `npm run test -- --run tests/unit/manage-workspace-components.test.tsx tests/integration/workbench-page.test.tsx`, `npm run typecheck`, `npm run lint`, `git diff --check`, `npm run build`. |
